Successfully implemented keypad on DE2 board.
Keypad works successfully with DE2 board. Several things need to be added though. Several signals need to be added to insure that the user cannot enter input during the period when the SRAM is writing. The keypad driver should also return an 16 bit value to the top entity for data inputted to the SRAM. It should also be able to return an 8 bit value for the SRAM  data address.